Implementare API GET pentru lista de medici dupa specialitate, ruleaza - NU este testat !
+ Alte modificari prin cod.

+++ src/main/java/com/example/dizertatie/entities/Medic.java
@@ -37,7 +37,7 @@ public class Medic {
     private String telefon;
 
     @Column(name = "COD_PARAFA")
-    private Integer codParafa;
+    private Long codParafa;
 
     @Column(name = "PASSWORD")
     private String parola;
